fixes based on the unit tests

+++ backend/src/seeds/graphql/queries.ts
@@ -301,7 +301,7 @@ export const searchAdminUsers = gql`
 `
 
 export const listContributionMessages = gql`
-  query ($contributionId: Float!, $pageSize: Int = 25, $currentPage: Int = 1, $order: Order = ASC) {
+  query ($contributionId: Int!, $pageSize: Int = 25, $currentPage: Int = 1, $order: Order = ASC) {
     listContributionMessages(
       contributionId: $contributionId
       pageSize: $pageSize
